feat(config): 更新config的配置
